requirements.txt
fastapi[all]
psycopg2
python-jose
# fastapi-sessions